input without heartbeat works
