+++ terminal_app/Cargo.toml
@@ -9,7 +9,7 @@ edition = "2021"
 
 zmq = "0.10.0"
 mac_address = "1.1.4"
-async-recursion = "1.0.0"
+threadpool = "1.8.1"
 futures = { version = "0.3.25" }
 futures-timer = { version ="3.0" }
 async-std = { version = "1.10", features = [ "attributes" ] }
