+++ terminal_app/src/services/mod.rs
@@ -1,1 +1,4 @@
 pub mod cache;
+pub mod input;
+pub mod get_mac;
+pub mod heartbeat;
